Final pre-presentation pushes

+++ covid_testing_site/expressEmployeeLogin/app.js
@@ -16,7 +16,7 @@ dotenv.config({ path: __dirname + "/.env" });
 // var indexRouter = require('./routes/index');
 // var usersRouter = require('./routes/users');
 var employeeLogin = require("./routes/employeeLogin");
-var collectLabLogin = require("./routes/collectLabLogin");
+var labTech = require("./routes/labTech");
 var labHomeButtons = require("./routes/labHomeButtons");
 const e = require("express");
 // var labLogin = require("./routes/labLogin");
@@ -59,7 +59,7 @@ pool.connect((error) => {
   }
 });
 
-console.log("hmmm");
+//console.log("hmmm");
 //pool.query(`select * from ${table}`)
 //    .then(res => {
 //      console.log(JSON.stringify(res));
@@ -80,13 +80,13 @@ app.use(express.static(path.join(__dirname, "/styles")));
 // app.use('/', indexRouter);
 // app.use('/blah', usersRouter);
 app.use("/employeeLogin", employeeLogin);
-app.use("/collectLabLogin", collectLabLogin);
+app.use("/labTech", labTech);
 app.use("/labHomeButtons", labHomeButtons);
 // app.use('/labLogin', labLogin);
 app.use("/employeeHome", (req, res) => {
   let query = url.parse(req.url, true).query;
-  console.log(query.email);
-  console.log(query.password);
+  //console.log(query.email);
+  //console.log(query.password);
   //if(query.email!=null){
   //if(query.email!='*'){
   //if(query.password!=null){
@@ -184,12 +184,34 @@ app.use("/poolMapping", (req, res) => {
   });
 });
 
-app.use("/addPool", (req, res) => {
+const query1 = util.promisify(pool.query).bind(pool);
+newPool = async (poolBarcode) => {
+  try {
+    const pool = await query1(
+      `SELECT * FROM covid.Pool WHERE poolBarcode = "${poolBarcode}";`
+    );
+
+    if (!pool.length) {
+      const newPool = await query1(
+        `INSERT INTO Pool VALUES ("${poolBarcode}");`
+      );
+      console.log("Added new pool: " + newPool);
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+app.use("/addPool", async (req, res) => {
   let query = url.parse(req.url, true).query;
   console.log(query);
   let poolBarcode = query.poolBarcode;
   let testBarCodes = query.barcode;
+  console.log("Function");
+  await newPool(poolBarcode);
+  console.log("After function");
   if (typeof testBarCodes === "string") {
+    console.log("String")
     pool.query(
       `INSERT INTO PoolMap (testBarCode, poolBarcode) 
         values ("${testBarCodes}","${poolBarcode}")`
@@ -197,6 +219,7 @@ app.use("/addPool", (req, res) => {
     res.json({ valid: "0" });
   } else {
     for (var i = 0; i < testBarCodes.length; i++) {
+      console.log("array")
       pool.query(
         `INSERT INTO PoolMap (testBarCode, poolBarcode) 
     values ("${testBarCodes[i]}","${poolBarcode}")`
@@ -280,22 +303,6 @@ app.use("/editDeletePool", (req, res) => {
   }
 });
 
-// app.use("/editPool", (req, res) => {
-//   let poolBarcode = req.body.poolBarcode;
-//   let testBarCodes = req.body.testBarCodes;
-//   pool.query(
-//     `DELETE FROM PoolMap poolBarcode
-//   WHERE poolBarcode = "${poolBarcode}";`);
-//   for(var i = 0; i < testBarCodes.length; i++)
-//   {
-//     pool.query(
-//       `DELETE FROM PoolMap testBarCode
-//     WHERE testBarCode = "${testBarCodes[i]}";`);
-//   }
-//   var send = {pool: poolBarcode, test: testBarCodes};
-//   res.send(send);
-// });
-
 app.use("/wellTest", (req, res) => {
   let query = url.parse(req.url, true).query;
   console.log(query);
@@ -308,20 +315,36 @@ app.use("/wellTest", (req, res) => {
   );
 });
 
-app.use("/addWell", (req, res) => {
+const query2 = util.promisify(pool.query).bind(pool);
+newWell = async (wellBarcode) => {
+  try {
+    const well = await query2(
+      `SELECT * FROM covid.Well WHERE wellBarcode = "${wellBarcode}";`
+    );
+
+    if (!well.length) {
+      const newWell = await query2(
+        `INSERT INTO Well VALUES ("${wellBarcode}");`
+      );
+      console.log("New Well: " + newWell);
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+app.use("/addWell", async (req, res) => {
   let query = url.parse(req.url, true).query;
-  console.log(query.wellBarcode);
-  console.log(query.action);
-  //addwell check for action?
   console.log(query.result);
   if (query.result === "In+Progress") {
     query.result = "In Progress";
   }
+  await newWell(query.wellBarcode);
   pool.query(
     `INSERT INTO WellTesting (poolBarcode, wellBarcode, testingStartTime, testingEndTime, result) 
   values ("${query.poolBarcode}","${query.wellBarcode}","2020-11-18 11:47:30", "2020-11-25 11:47:30","${query.result}")`,
     (error, data, fields) => {
-      meep = blah;
+      meep = data;
       console.log(meep);
       res.send(meep);
     }
@@ -356,7 +379,7 @@ app.use("/editOrDelete", (req, res) => {
   AND poolBarcode='${x[1]}'
   AND result='${x[2]}'`,
       (error, data, fields) => {
-        meep = blah;
+        meep = data;
         console.log(meep);
         if (typeof query.WellPoolResult == "string") {
           pool.query(`DELETE FROM WellTesting
@@ -405,8 +428,8 @@ app.use("/testCollection", (req, res) => {
 
 app.use("/addTest", (req, res) => {
   let query = url.parse(req.url, true).query;
-  console.log(query.employeeID);
-  console.log(query.testBarcode);
+  //console.log(query.employeeID);
+  //console.log(query.testBarcode);
   var x = new Date();
   var y = "";
   y +=
@@ -424,7 +447,7 @@ app.use("/addTest", (req, res) => {
   //addwell check for action?
   pool.query(
     `INSERT INTO EmployeeTest (testBarcode, employeeID, collectionTime, collectedBy) 
-  values ("${query.testBarcode}","${query.employeeID}","${y}","L1")`,
+  values ("${query.testBarcode}","${query.employeeID}","${y}","L27804")`,
     (error, blah, fields) => {
       meep = blah;
       console.log(meep);
@@ -433,9 +456,25 @@ app.use("/addTest", (req, res) => {
   );
 });
 
-app.use("/deleteTest", (req, res) => {
-  let query = url.parse(req.url, true).query;
+const query3 = util.promisify(pool.query).bind(pool);
+canDelete = async (testBarcode) => {
+  try {
+    const testTable = await query3(
+      `SELECT * FROM covid.PoolMap WHERE testBarCode = "${testBarcode}";`
+    );
 
+    if (!testTable.length) {
+      return true;
+    }
+
+    return false;
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+app.use("/deleteTest", async (req, res) => {
+  let query = url.parse(req.url, true).query;
   var x = query.IdBarcode;
   console.log(x);
   if (x == null) {
@@ -457,15 +496,21 @@ app.use("/deleteTest", (req, res) => {
     for (i = 0; i < x.length; i++) {
       y = x[i].split(",");
       console.log(y);
-      pool.query(`DELETE FROM EmployeeTest
+      var checkDelete = await canDelete(y[1]);
+      if (checkDelete) {
+        pool.query(`DELETE FROM EmployeeTest
     WHERE employeeID='${y[0]}'
     AND testBarcode='${y[1]}'`);
+      }
     }
   } else {
     console.log("asdsad");
-    pool.query(`DELETE FROM EmployeeTest
+    var checkDelete = await canDelete(x[1]);
+    if (checkDelete) {
+      pool.query(`DELETE FROM EmployeeTest
     WHERE employeeID='${x[0]}'
     AND testBarcode='${x[1]}'`);
+    }
   }
 
   res.send([]);
